requirements.txt
fastapi
uvicorn[standard]
python-multipart
chromadb
python-dotenv
langchain
langchain-chroma
langchain-classic
langchain-google-genai
langchain-huggingface
sentence-transformers
pymupdf
